requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langchain==0.3.27
langgraph==0.6.5
langchain-openai==0.3.30
langchain-community==0.3.27
langchain-text-splitters==0.3.9
langchain-experimental==0.3.4
faiss-cpu==1.12.0
crawl4ai==0.7.3
beautifulsoup4==4.13.5
requests==2.32.5
python-dotenv==1.1.1
tiktoken==0.11.0
